# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Evolve Performance e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List active products
      parameters:
        - name: collection
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /products/{slug}:
    get:
      operationId: getProductBySlug
      tags: [products]
      summary: Get a product by slug with variants and images
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /checkout/session:
    post:
      operationId: createCheckoutSession
      tags: [checkout]
      summary: Create a Stripe checkout session with server-computed totals
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutSessionInput"
      responses:
        "200":
          description: Stripe checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSessionResponse"
        "400":
          description: Invalid cart or discount
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /checkout/validate-discount:
    post:
      operationId: validateDiscountCode
      tags: [checkout]
      summary: Validate a discount code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DiscountValidateInput"
      responses:
        "200":
          description: Discount details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscountValidateResponse"
        "404":
          description: Invalid or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders/lookup:
    get:
      operationId: lookupOrder
      tags: [orders]
      summary: Look up order by order number and email
      parameters:
        - name: orderNumber
          in: query
          required: true
          schema:
            type: string
        - name: email
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Tracking ──────────────────────────────────────────────────────────────────
...
  /account/orders:
    get:
      operationId: listAccountOrders
      tags: [account]
      summary: List orders for the authenticated customer
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Customer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter (records consent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterSubscribeInput"
      responses:
        "200":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterSubscribeResponse"
  /newsletter/unsubscribe:
    post:
      operationId: unsubscribeNewsletter
      tags: [newsletter]
      summary: Unsubscribe from newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterUnsubscribeInput"
      responses:
        "200":
          description: Unsubscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsletterUnsubscribeResponse"

  # ── Content ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Collection:
      type: object
      required: [id, slug, name, description, imageUrl, productCount]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        productCount:
          type: integer
...
    ProductCard:
      type: object
      required: [id, slug, name, priceMin, priceMax, primaryImageUrl, collectionName, status, stockStatus]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        theme:
          type: ["string", "null"]
        priceMin:
          type: integer
        priceMax:
          type: integer
        primaryImageUrl:
          type: ["string", "null"]
        collectionName:
          type: ["string", "null"]
        status:
          type: string
        stockStatus:
          type: string
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
    ProductListResponse:
      type: object
      required: [products, total]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/ProductCard"
        total:
          type: integer
    ProductDetail:
      type: object
      required: [id, slug, name, description, variants, images, collections]
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        theme:
          type: ["string", "null"]
        description:
          type: string
        materials:
          type: ["string", "null"]
        benefits:
          type: ["string", "null"]
        shippingInfo:
          type: ["string", "null"]
        returnsInfo:
          type: ["string", "null"]
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        collections:
          type: array
          items:
            $ref: "#/components/schemas/Collection"
        averageRating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        weightGrams:
          type: ["integer", "null"]
        dimensionsCm:
          type: ["string", "null"]
        hsCode:
          type: ["string", "null"]
        countryOfOrigin:
          type: ["string", "null"]
    ProductVariant:
      type: object
      required: [id, sku, priceInCents, stockQuantity, availableQuantity]
      properties:
        id:
          type: integer
        sku:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        priceInCents:
          type: integer
        compareAtPriceInCents:
          type: ["integer", "null"]
        stockQuantity:
          type: integer
        availableQuantity:
          type: integer
        imageUrl:
          type: ["string", "null"]
    ProductImage:
      type: object
      required: [id, url, altText, position]
      properties:
        id:
          type: integer
        url:
          type: string
        altText:
          type: ["string", "null"]
        position:
          type: integer

    # ── Cart ─────────────────────────────────────────────────────────────────
...
    CheckoutSessionInput:
      type: object
      required: [sessionId, successUrl, cancelUrl]
      properties:
        sessionId:
          type: string
        clerkUserId:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        discountCode:
          type: ["string", "null"]
        shippingZoneRateId:
          type: ["integer", "null"]
        successUrl:
          type: string
        cancelUrl:
          type: string
    CheckoutSessionResponse:
      type: object
      required: [checkoutUrl, subtotalInCents, shippingInCents, taxInCents, totalInCents]
      properties:
        checkoutUrl:
          type: string
        subtotalInCents:
          type: integer
        shippingInCents:
          type: integer
        taxInCents:
          type: integer
        discountInCents:
          type: integer
        totalInCents:
          type: integer
...
    DiscountValidateInput:
      type: object
      required: [code, sessionId]
      properties:
        code:
          type: string
        sessionId:
          type: string
    DiscountValidateResponse:
      type: object
      required: [valid, code]
      properties:
        valid:
          type: boolean
        code:
          type: string
        discountType:
          type: ["string", "null"]
        value:
          type: ["integer", "null"]
        discountInCents:
          type: ["integer", "null"]
        error:
          type: ["string", "null"]

    # ── Orders ────────────────────────────────────────────────────────────────
    OrderSummary:
      type: object
      required: [id, orderNumber, status, totalInCents, currency, itemCount, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        totalInCents:
          type: integer
        currency:
          type: string
        itemCount:
          type: integer
        trackingNumber:
          type: ["string", "null"]
        carrier:
          type: ["string", "null"]
        createdAt:
          type: string
    OrderDetail:
      type: object
      required: [id, orderNumber, status, items, subtotalInCents, shippingInCents, taxInCents, discountInCents, totalInCents, currency, shippingAddress, createdAt]
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        subtotalInCents:
          type: integer
        shippingInCents:
          type: integer
        taxInCents:
          type: integer
        discountInCents:
          type: integer
        totalInCents:
          type: integer
        currency:
          type: string
        shippingAddress:
          $ref: "#/components/schemas/AddressSnapshot"
        trackingNumber:
          type: ["string", "null"]
        carrier:
          type: ["string", "null"]
        trackingUrl:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        stripeRiskLevel:
          type: ["string", "null"]
        requiresManualReview:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderItem:
      type: object
      required: [id, productName, variantSku, quantity, unitPriceInCents, totalPriceInCents]
      properties:
        id:
          type: integer
        productName:
          type: string
        productSlug:
          type: ["string", "null"]
        variantSku:
          type: string
        size:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        quantity:
          type: integer
        unitPriceInCents:
          type: integer
        totalPriceInCents:
          type: integer
        imageUrl:
          type: ["string", "null"]

    # ── Tracking ─────────────────────────────────────────────────────────────
...
    AddressSnapshot:
      type: object
      required: [name, line1, city, countryCode]
      properties:
        name:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        countryCode:
          type: string
...
    NewsletterSubscribeInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        firstName:
          type: string
        source:
          type: string
    NewsletterSubscribeResponse:
      type: object
      required: [subscribed]
      properties:
        subscribed:
          type: boolean
    NewsletterUnsubscribeInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
        token:
          type: ["string", "null"]
    NewsletterUnsubscribeResponse:
      type: object
      required: [unsubscribed]
      properties:
        unsubscribed:
          type: boolean

    # ── Content ───────────────────────────────────────────────────────────────
...