# lib/api-zod/src/generated/types/addressSnapshot.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Evolve Performance e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AddressSnapshot {
  name: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  /** @nullable */
  state?: string | null;
  /** @nullable */
  postalCode?: string | null;
  countryCode: string;
}